Major refactor of functions to class
