+++ geoapps/__init__.py
@@ -1,1 +1,13 @@
-from . import simpegEM1D, simpegPF, processing, utils
+from . import (
+    simpegEM1D,
+    simpegPF,
+    selection,
+    utils,
+    processing,
+    simulation,
+    export,
+    plotting,
+)
+
+# from .plotting import plot_plan_data_selection
+# from .selection import ObjectDataSelection
