+++ geoapps/plotting.py
@@ -1,11 +1,283 @@
+import re
 import ipywidgets as widgets
 import matplotlib.colors as colors
 import matplotlib.pyplot as plt
 import numpy as np
 from geoh5py.objects import Curve, Grid2D, Points, Surface
 from geoh5py.workspace import Workspace
-from ipywidgets.widgets import Dropdown, HBox, VBox
-from geoapps.utils import filter_xy
+from ipywidgets import (
+    Dropdown,
+    ColorPicker,
+    SelectMultiple,
+    Text,
+    IntSlider,
+    Checkbox,
+    FloatSlider,
+    FloatText,
+    VBox,
+    HBox,
+    ToggleButton,
+    ToggleButtons,
+    interactive_output,
+    FloatLogSlider,
+    Label,
+    Layout,
+    RadioButtons,
+)
+from geoapps.base import Widget
+from geoapps.utils import filter_xy, rotate_xy, format_labels
+from geoapps.selection import ObjectDataSelection
+
+
+class PlotSelection2D(Widget):
+    """
+    Application for selecting data in 2D plan map view
+    """
+
+    def __init__(self, h5file, **kwargs):
+
+        self.workspace = Workspace(h5file)
+
+        self.selection = ObjectDataSelection(h5file)
+        self.objects = self.selection.objects
+
+        self.center_x = FloatSlider(
+            min=-100, max=100, steps=10, description="Easting", continuous_update=False,
+        )
+        self.center_y = FloatSlider(
+            min=-100,
+            max=100,
+            steps=10,
+            description="Northing",
+            continuous_update=False,
+            orientation="vertical",
+        )
+        self.azimuth = FloatSlider(
+            min=-90,
+            max=90,
+            value=0,
+            steps=5,
+            description="Orientation",
+            continuous_update=False,
+        )
+        self.width_x = FloatSlider(
+            min=0,
+            max=100,
+            steps=10,
+            value=1000,
+            description="Width",
+            continuous_update=False,
+        )
+        self.width_y = FloatSlider(
+            min=0,
+            max=100,
+            steps=10,
+            value=1000,
+            description="Height",
+            continuous_update=False,
+            orientation="vertical",
+        )
+        self.resolution = FloatText(
+            description="Grid Resolution (m)", style={"description_width": "initial"},
+        )
+        self.data_count = Label("Data Count: 0", tooltip="Keep <1500 for speed")
+        self.zoom_extent = ToggleButton(
+            value=True,
+            description="Zoom on selection",
+            tooltip="Keep plot extent on selection",
+            icon="check",
+        )
+
+        if "contours" in kwargs.keys():
+            contours = kwargs["contours"]
+        else:
+            contours = ""
+
+        self.contours = widgets.Text(
+            value=contours,
+            description="Contours",
+            disabled=False,
+            continuous_update=False,
+        )
+
+        def set_bounding_box(_):
+            self.set_bounding_box()
+
+        self.selection.objects.observe(set_bounding_box)
+        self.pause = False
+
+        def plot_selection(
+            data_name,
+            resolution,
+            center_x,
+            center_y,
+            width_x,
+            width_y,
+            azimuth,
+            zoom_extent,
+            contours,
+        ):
+
+            self.plot_selection(
+                data_name,
+                resolution,
+                center_x,
+                center_y,
+                width_x,
+                width_y,
+                azimuth,
+                zoom_extent,
+                contours,
+            )
+
+        self.window_plot = widgets.interactive_output(
+            plot_selection,
+            {
+                "data_name": self.selection.data,
+                "resolution": self.resolution,
+                "center_x": self.center_x,
+                "center_y": self.center_y,
+                "width_x": self.width_x,
+                "width_y": self.width_y,
+                "azimuth": self.azimuth,
+                "zoom_extent": self.zoom_extent,
+                "contours": self.contours,
+            },
+        )
+        self.widget = VBox(
+            [
+                VBox([self.selection.widget, self.resolution, self.data_count,]),
+                HBox(
+                    [
+                        self.center_y,
+                        self.width_y,
+                        VBox(
+                            [
+                                self.width_x,
+                                self.center_x,
+                                self.window_plot,
+                                self.azimuth,
+                                self.zoom_extent,
+                            ]
+                        ),
+                    ],
+                    layout=Layout(align_items="center"),
+                ),
+            ]
+        )
+        self.figure = None
+        self.axis = None
+        self.indices = None
+
+        super().__init__(**kwargs)
+
+    def plot_selection(
+        self,
+        data_name,
+        resolution,
+        center_x,
+        center_y,
+        width_x,
+        width_y,
+        azimuth,
+        zoom_extent,
+        contours,
+    ):
+
+        if self.pause:
+            return
+
+        # Parse the contours string
+        if contours != "":
+            vals = re.split(",", contours)
+            cntrs = []
+            for val in vals:
+                if ":" in val:
+                    param = np.asarray(re.split(":", val), dtype="int")
+                    if len(param) == 2:
+                        cntrs += [np.arange(param[0], param[1])]
+                    else:
+                        cntrs += [np.arange(param[0], param[1], param[2])]
+                else:
+                    cntrs += [np.float(val)]
+            contours = np.unique(np.sort(np.hstack(cntrs)))
+        else:
+            contours = None
+
+        entity, data_obj = self.selection.get_selected_entities()
+
+        if isinstance(entity, (Grid2D, Surface, Points, Curve)):
+
+            self.figure = plt.figure(figsize=(10, 10))
+
+            self.axis = plt.subplot()
+            corners = np.r_[
+                np.c_[-1.0, -1.0],
+                np.c_[-1.0, 1.0],
+                np.c_[1.0, 1.0],
+                np.c_[1.0, -1.0],
+                np.c_[-1.0, -1.0],
+            ]
+            corners[:, 0] *= width_x / 2
+            corners[:, 1] *= width_y / 2
+            corners = rotate_xy(corners, [0, 0], -azimuth)
+            self.axis.plot(corners[:, 0] + center_x, corners[:, 1] + center_y, "k")
+
+            _, _, ind_filter, _, contour_set = plot_plan_data_selection(
+                entity,
+                data_obj,
+                **{
+                    "ax": self.axis,
+                    "resolution": resolution,
+                    "window": {
+                        "center": [center_x, center_y],
+                        "size": [width_x, width_y],
+                        "azimuth": azimuth,
+                    },
+                    "zoom_extent": zoom_extent,
+                    "resize": True,
+                    "contours": contours,
+                },
+            )
+            self.indices = ind_filter
+            self.contours.contour_set = contour_set
+            self.data_count.value = f"Data Count: {ind_filter.sum()}"
+
+    def set_bounding_box(self):
+        # Fetch vertices in the project
+        lim_x = [1e8, -1e8]
+        lim_y = [1e8, -1e8]
+
+        obj, _ = self.selection.get_selected_entities()
+        if isinstance(obj, Grid2D):
+            lim_x[0], lim_x[1] = obj.centroids[:, 0].min(), obj.centroids[:, 0].max()
+            lim_y[0], lim_y[1] = obj.centroids[:, 1].min(), obj.centroids[:, 1].max()
+        elif isinstance(obj, (Points, Curve, Surface)):
+            lim_x[0], lim_x[1] = obj.vertices[:, 0].min(), obj.vertices[:, 0].max()
+            lim_y[0], lim_y[1] = obj.vertices[:, 1].min(), obj.vertices[:, 1].max()
+        else:
+            return
+
+        self.pause = True
+        self.center_x.min = -np.inf
+        self.center_x.max = lim_x[1]
+        self.center_x.value = np.mean(lim_x)
+        self.center_x.min = lim_x[0]
+
+        self.center_y.min = -np.inf
+        self.center_y.max = lim_y[1]
+        self.center_y.value = np.mean(lim_y)
+        self.center_y.min = lim_y[0]
+
+        self.width_x.max = lim_x[1] - lim_x[0]
+        self.width_x.value = self.width_x.max / 2.0
+        self.width_x.min = 0
+
+        self.width_y.max = lim_y[1] - lim_y[0]
+        self.width_y.min = 0
+
+        self.pause = False
+        self.width_y.value = self.width_y.max / 2.0
 
 
 def plot_profile_data_selection(
@@ -97,8 +369,6 @@ def plot_profile_data_selection(
                     aspect="auto",
                 )
 
-    # ax.set_position([pos.x0, pos.y0, pos.width*2., pos.height])
-    # ax.set_aspect(1)
     return ax, threshold
 
 
@@ -118,12 +388,15 @@ def plot_plan_data_selection(entity, data, **kwargs):
         indices = None
 
     line_selection = None
+    contour_set = None
     values = None
     ax = None
+    out = None
+
     if isinstance(getattr(data, "values", None), np.ndarray):
         if not isinstance(data.values[0], str):
             values = data.values.copy()
-            values[np.abs(values) < 1e-18] = np.nan
+            values[(values > 1e-18) * (values < 2e-18)] = np.nan
             values[values == -99999] = np.nan
 
     color_norm = None
@@ -134,138 +407,126 @@ def plot_plan_data_selection(entity, data, **kwargs):
     if "window" in kwargs.keys():
         window = kwargs["window"]
 
-    if values is not None:
+    if data is not None and data.entity_type.color_map is not None:
+        new_cmap = data.entity_type.color_map.values
+        map_vals = new_cmap["Value"].copy()
+        cmap = colors.ListedColormap(
+            np.c_[
+                new_cmap["Red"] / 255, new_cmap["Green"] / 255, new_cmap["Blue"] / 255,
+            ]
+        )
+        color_norm = colors.BoundaryNorm(map_vals, cmap.N)
+    else:
+        cmap = "Spectral_r"
 
-        if data.entity_type.color_map is not None:
-            new_cmap = data.entity_type.color_map.values
-            map_vals = new_cmap["Value"].copy()
-            cmap = colors.ListedColormap(
-                np.c_[
-                    new_cmap["Red"] / 255,
-                    new_cmap["Green"] / 255,
-                    new_cmap["Blue"] / 255,
-                ]
-            )
-            color_norm = colors.BoundaryNorm(map_vals, cmap.N)
-        else:
-            cmap = "Spectral_r"
+    if "ax" not in kwargs.keys():
+        plt.figure(figsize=(8, 8))
+        ax = plt.subplot()
+    else:
+        ax = kwargs["ax"]
 
-        if "ax" not in kwargs.keys():
-            plt.figure(figsize=(8, 8))
-            ax = plt.subplot()
-        else:
-            ax = kwargs["ax"]
+    if isinstance(entity, Grid2D) and values is not None:
+        x = entity.centroids[:, 0].reshape(entity.shape, order="F")
+        y = entity.centroids[:, 1].reshape(entity.shape, order="F")
+        indices = filter_xy(x, y, resolution, window=window)
 
-        if np.all(np.isnan(values)):
-            return (
-                ax,
-                np.zeros_like(values, dtype="bool"),
-                np.zeros_like(values, dtype="bool"),
-            )
-        out = None
-        if isinstance(entity, Grid2D) and values is not None:
-            x = entity.centroids[:, 0].reshape(entity.shape, order="F")
-            y = entity.centroids[:, 1].reshape(entity.shape, order="F")
+        ind_x, ind_y = (
+            np.any(indices, axis=1),
+            np.any(indices, axis=0),
+        )
+
+        X = x[ind_x, :][:, ind_y]
+        Y = y[ind_x, :][:, ind_y]
+
+        if values is not None:
             values = values.reshape(entity.shape, order="F")
-            indices = filter_xy(x, y, resolution, window=window)
             values[indices == False] = np.nan
+            values = values[ind_x, :][:, ind_y]
 
-            ind_x, ind_y = (
-                np.any(indices, axis=1),
-                np.any(indices, axis=0),
+        if np.any(values):
+            out = ax.pcolormesh(
+                X, Y, values, cmap=cmap, norm=color_norm, shading="auto"
             )
 
-            X = x[ind_x, :][:, ind_y]
-            Y = y[ind_x, :][:, ind_y]
-            vals = values[ind_x, :][:, ind_y]
+        if "contours" in kwargs.keys() and np.any(values):
+            contour_set = ax.contour(
+                X, Y, values, levels=kwargs["contours"], colors="k", linewidths=1.0
+            )
 
-            if np.any(vals):
-                out = ax.pcolormesh(
-                    X, Y, vals, cmap=cmap, norm=color_norm, shading="auto"
-                )
+    elif isinstance(entity, (Points, Curve, Surface)):
+        if indices is None:
+            indices = filter_xy(
+                entity.vertices[:, 0], entity.vertices[:, 1], resolution, window=window,
+            )
 
-            if "contours" in kwargs.keys():
-                ax.contour(
-                    x, y, values, levels=kwargs["contours"], colors="k", linewidths=1.0
-                )
-        elif (
-            isinstance(entity, Points)
-            or isinstance(entity, Curve)
-            or isinstance(entity, Surface)
-        ):
-
-            if indices is None:
-                indices = filter_xy(
-                    entity.vertices[:, 0],
-                    entity.vertices[:, 1],
-                    resolution,
-                    window=window,
-                )
-
-            x, y = entity.vertices[indices, 0], entity.vertices[indices, 1]
+        x, y = entity.vertices[:, 0], entity.vertices[:, 1]
+        X, Y = x[indices], y[indices]
+        if values is not None:
             values = values[indices]
 
-            if "marker_size" not in kwargs.keys():
-                marker_size = 5
-            else:
-                marker_size = kwargs["marker_size"]
-
-            out = ax.scatter(x, y, marker_size, values, cmap=cmap, norm=color_norm)
-
-            if "contours" in kwargs.keys():
-                ind = ~np.isnan(values)
-                ax.tricontour(
-                    x[ind],
-                    y[ind],
-                    values[ind],
-                    levels=kwargs["contours"],
-                    colors="k",
-                    linewidths=1.0,
-                )
-
+        if "marker_size" not in kwargs.keys():
+            marker_size = 5
         else:
-            print(
-                "Sorry, 'plot=True' option only implemented for Grid2D, Points and Curve objects"
-            )
-            out = None
+            marker_size = kwargs["marker_size"]
 
-        if "zoom_extent" in kwargs.keys() and kwargs["zoom_extent"]:
+        out = ax.scatter(X, Y, marker_size, values, cmap=cmap, norm=color_norm)
+
+        if "contours" in kwargs.keys() and np.any(values):
             ind = ~np.isnan(values)
-            if ind.sum() > 0:
-                format_labels(x[ind], y[ind], ax)
-                ax.set_xlim([x[ind].min(), x[ind].max()])
-                ax.set_ylim([y[ind].min(), y[ind].max()])
-        else:
+            contour_set = ax.tricontour(
+                X[ind],
+                Y[ind],
+                values[ind],
+                levels=kwargs["contours"],
+                colors="k",
+                linewidths=1.0,
+            )
+
+    else:
+        print(
+            "Sorry, 'plot=True' option only implemented for Grid2D, Points, Surface and Curve objects"
+        )
+        out = None
+
+    if "zoom_extent" in kwargs.keys() and kwargs["zoom_extent"] and np.any(values):
+        ind = ~np.isnan(values.ravel())
+        x = X.ravel()[ind]
+        y = Y.ravel()[ind]
+        if ind.sum() > 0:
             format_labels(x, y, ax)
+            ax.set_xlim([x.min(), x.max()])
+            ax.set_ylim([y.min(), y.max()])
+    elif np.any(x) and np.any(y):
+        format_labels(x, y, ax)
+        ax.set_xlim([x.min(), x.max()])
+        ax.set_ylim([y.min(), y.max()])
 
-        if (
-            "colorbar" in kwargs.keys()
-            and values[~np.isnan(values)].min() != values[~np.isnan(values)].max()
-        ):
-            plt.colorbar(out, ax=ax)
+    if (
+        "colorbar" in kwargs.keys()
+        and values[~np.isnan(values)].min() != values[~np.isnan(values)].max()
+    ):
+        plt.colorbar(out, ax=ax)
 
-        line_selection = np.zeros_like(indices, dtype=bool)
-        if "highlight_selection" in kwargs.keys():
-            for key, values in kwargs["highlight_selection"].items():
+    line_selection = np.zeros_like(indices, dtype=bool)
+    if "highlight_selection" in kwargs.keys():
+        for key, values in kwargs["highlight_selection"].items():
 
-                if not np.any(entity.get_data(key)):
-                    continue
+            if not np.any(entity.get_data(key)):
+                continue
 
-                for line in values:
-                    ind = np.where(entity.get_data(key)[0].values == line)[0]
-                    x, y, values = (
-                        locations[ind, 0],
-                        locations[ind, 1],
-                        entity.get_data(key)[0].values[ind],
-                    )
-                    ind_line = filter_xy(x, y, resolution, window=window)
-                    ax.scatter(
-                        x[ind_line], y[ind_line], marker_size * 2, "k", marker="+"
-                    )
+            for line in values:
+                ind = np.where(entity.get_data(key)[0].values == line)[0]
+                x, y, values = (
+                    locations[ind, 0],
+                    locations[ind, 1],
+                    entity.get_data(key)[0].values[ind],
+                )
+                ind_line = filter_xy(x, y, resolution, window=window)
+                ax.scatter(x[ind_line], y[ind_line], marker_size * 2, "k", marker="+")
 
-                    line_selection[ind[ind_line]] = True
+                line_selection[ind[ind_line]] = True
 
-    return ax, out, indices, line_selection
+    return ax, out, indices, line_selection, contour_set
 
 
 def plot_em_data_widget(h5file):
@@ -347,7 +608,7 @@ def plot_em_data_widget(h5file):
         lines.options = line_ids
         lines.value = value
 
-    objects = widgets.Dropdown(options=names, value=names[0], description="Object:",)
+    objects = Dropdown(options=names, value=names[0], description="Object:",)
 
     obj = get_parental_child(objects.value)[0]
 
@@ -364,9 +625,7 @@ def plot_em_data_widget(h5file):
 
     options = [pg.name for pg in obj.property_groups]
     options = [option for option in sorted(options)]
-    groups = widgets.SelectMultiple(
-        options=options, value=[options[0]], description="Data: ",
-    )
+    groups = SelectMultiple(options=options, value=[options[0]], description="Data: ",)
 
     if line_field.value is None:
         line_list = []
@@ -376,7 +635,7 @@ def plot_em_data_widget(h5file):
         line_list = np.unique(entity.get_data(line_field.value)[0].values)
         value = [line_list[0]]
 
-    lines = widgets.SelectMultiple(options=line_list, value=value, description="Data: ")
+    lines = SelectMultiple(options=line_list, value=value, description="Data: ")
 
     objects.observe(updateList, names="value")
 
@@ -384,7 +643,7 @@ def plot_em_data_widget(h5file):
         options=["linear", "symlog"], value="symlog", description="Scaling",
     )
 
-    threshold = widgets.FloatSlider(
+    threshold = FloatSlider(
         min=-16,
         max=-1,
         value=-12,
@@ -398,7 +657,7 @@ def plot_em_data_widget(h5file):
     layout = HBox(
         [
             apps,
-            widgets.interactive_output(
+            interactive_output(
                 plot_profiles,
                 {
                     "entity_name": objects,
@@ -412,23 +671,3 @@ def plot_em_data_widget(h5file):
         ]
     )
     return layout
-
-
-def format_labels(x, y, axs, labels=None, aspect="equal", tick_format="%i"):
-    if labels is None:
-        axs.set_ylabel("Northing (m)")
-        axs.set_xlabel("Easting (m)")
-    else:
-        axs.set_xlabel(labels[0])
-        axs.set_ylabel(labels[1])
-    xticks = np.linspace(x.min(), x.max(), 5)
-    yticks = np.linspace(y.min(), y.max(), 5)
-
-    axs.set_yticks(yticks)
-    axs.set_yticklabels(
-        [tick_format % y for y in yticks.tolist()], rotation=90, va="center"
-    )
-    axs.set_xticks(xticks)
-    axs.set_xticklabels([tick_format % x for x in xticks.tolist()], va="center")
-    axs.autoscale(tight=True)
-    axs.set_aspect(aspect)
