+++ geoapps/processing/processing.py
@@ -16,14 +16,13 @@ from scipy.interpolate import LinearNDInterpolator, interp1d
 from scipy.spatial import cKDTree, Delaunay
 from skimage.feature import canny
 from skimage.transform import probabilistic_hough_line
-from geoapps import utils
-from geoapps.plotting import format_labels
+from geoapps.utils import format_labels
 from geoapps.inversion import TopographyOptions
 from geoapps.selection import (
-    object_data_selection_widget,
-    plot_plan_data_selection,
+    ObjectDataSelection,
     LineOptions,
 )
+from geoapps.plotting import plot_plan_data_selection
 from geoapps.utils import (
     filter_xy,
     export_grid_2_geotiff,
@@ -36,10 +35,10 @@ from geoapps.utils import (
 
 def calculator(h5file):
     w_s = Workspace(h5file)
-    objects, data = object_data_selection_widget(h5file, select_multiple=False)
-    _, store = object_data_selection_widget(
-        h5file, objects=objects, select_multiple=False
-    )
+    selection = ObjectDataSelection(h5file, select_multiple=False)
+    _, store = ObjectDataSelection(
+        h5file, objects=selection.objects, select_multiple=False
+    ).data
     store.description = "Assign result to: "
     store.style = {"description_width": "initial"}
     use = widgets.ToggleButton(description=">> Add >>")
@@ -54,14 +53,14 @@ def calculator(h5file):
 
     def evaluate(var):
         vals = eval(equation.value)
-        obj = w_s.get_entity(objects.value)[0]
+        obj = w_s.get_entity(selection.objects.value)[0]
         obj.get_data(store.value)[0].values = vals
         print(vals)
         w_s.finalize()
 
     def click_add(_):
         if add.value:
-            obj = w_s.get_entity(objects.value)[0]
+            obj = w_s.get_entity(selection.objects.value)[0]
 
             if getattr(obj, "vertices", None) is not None:
                 new_data = obj.add_data(
@@ -72,7 +71,7 @@ def calculator(h5file):
                     {channel.value: {"values": numpy.zeros(obj.n_cells)}}
                 )
 
-            data.options = obj.get_data_list()
+            selection.data.options = obj.get_data_list()
             store.options = [new_data.name]
             store.value = new_data.name
 
@@ -80,10 +79,10 @@ def calculator(h5file):
 
     def click_use(_):
         if use.value:
-            name = objects.value + "." + data.value
+            name = selection.objects.value + "." + selection.data.value
             if name not in var.keys():
-                obj = w_s.get_entity(objects.value)[0]
-                var[name] = obj.get_data(data.value)[0].values
+                obj = w_s.get_entity(selection.objects.value)[0]
+                var[name] = obj.get_data(selection.data.value)[0].values
 
             equation.value = equation.value + "var['" + name + "']"
             use.value = False
@@ -99,8 +98,8 @@ def calculator(h5file):
 
     return VBox(
         [
-            objects,
-            HBox([use, data]),
+            selection.objects,
+            HBox([use, selection.data]),
             VBox(
                 [equation, HBox([add, channel]), store, compute],
                 layout=Layout(width="100%"),
@@ -117,18 +116,18 @@ def cdi_curve_2_surface(h5file):
 
     workspace = Workspace(h5file)
 
-    objects, models = object_data_selection_widget(h5file, add_groups=True)
-    models.description = "Model fields: "
+    selection = ObjectDataSelection(h5file, add_groups=True)
+    selection.data.description = "Model fields: "
 
-    _, line_channel = object_data_selection_widget(
-        h5file, add_groups=True, objects=objects, find_value=["line"]
+    _, line_channel = ObjectDataSelection(
+        h5file, add_groups=True, objects=selection.objects, find_value=["line"]
     )
     line_channel.description = "Line field:"
 
     topo_options = TopographyOptions(h5file)
 
-    _, elevations = object_data_selection_widget(
-        h5file, add_groups=True, objects=objects
+    _, elevations = ObjectDataSelection(
+        h5file, add_groups=True, objects=selection.objects
     )
     elevations.description = "Elevations:"
 
@@ -172,8 +171,8 @@ def cdi_curve_2_surface(h5file):
     def convert_trigger(_):
 
         if convert.value:
-            if workspace.get_entity(objects.value):
-                curve = workspace.get_entity(objects.value)[0]
+            if workspace.get_entity(selection.objects.value):
+                curve = workspace.get_entity(selection.objects.value)[0]
                 convert.value = False
             else:
                 convert.button_style = "warning"
@@ -246,7 +245,7 @@ def cdi_curve_2_surface(h5file):
                 for ind, (z_prop, m_prop) in enumerate(
                     zip(
                         curve.get_property_group(elevations.value).properties,
-                        curve.get_property_group(models.value).properties,
+                        curve.get_property_group(selection.data.value).properties,
                     )
                 ):
                     nZ += 1
@@ -351,7 +350,7 @@ def cdi_curve_2_surface(h5file):
 
             surface.add_data(
                 {
-                    models.value: {"values": numpy.hstack(model)},
+                    selection.data.value: {"values": numpy.hstack(model)},
                     "Line": {"values": numpy.hstack(line_ids)},
                 }
             )
@@ -364,9 +363,8 @@ def cdi_curve_2_surface(h5file):
                 [
                     widgets.VBox(
                         [
-                            objects,
+                            selection.widget,
                             line_channel,
-                            models,
                             depth_panel,
                             widgets.Label("Triangulation"),
                             max_depth,
@@ -618,11 +616,12 @@ def edge_detection_widget(
 
     workspace = Workspace(h5file)
 
-    objects, data_obj = object_data_selection_widget(h5file)
+    selection = ObjectDataSelection(h5file)
 
     center_x = widgets.FloatSlider(
         min=-100, max=100, steps=10, description="Easting", continuous_update=False,
     )
+
     center_y = widgets.FloatSlider(
         min=-100,
         max=100,
@@ -663,19 +662,18 @@ def edge_detection_widget(
         description="Grid Resolution (m)",
         style={"description_width": "initial"},
     )
-
-    window_size = widgets.IntText(
-        value=64,
-        description="Window size (pixels)",
-        style={"description_width": "initial"},
-    )
-
     data_count = Label("Data Count: 0", tooltip="Keep <1500 for speed")
     zoom_extent = widgets.ToggleButton(
         value=True,
         description="Zoom on selection",
         tooltip="Keep plot extent on selection",
         icon="check",
+    )
+
+    window_size = widgets.IntText(
+        value=64,
+        description="Window size (pixels)",
+        style={"description_width": "initial"},
     )
     export = widgets.ToggleButton(
         value=False,
@@ -700,7 +698,6 @@ def edge_detection_widget(
         description="Sigma",
         style={"description_width": "initial"},
     )
-
     line_length = widgets.IntSlider(
         min=1.0,
         max=100.0,
@@ -710,7 +707,6 @@ def edge_detection_widget(
         description="Line Length",
         style={"description_width": "initial"},
     )
-
     line_gap = widgets.IntSlider(
         min=1.0,
         max=100.0,
@@ -720,7 +716,6 @@ def edge_detection_widget(
         description="Line Gap",
         style={"description_width": "initial"},
     )
-
     threshold = widgets.IntSlider(
         min=1.0,
         max=100.0,
@@ -747,10 +742,10 @@ def edge_detection_widget(
         lim_x = [1e8, -1e8]
         lim_y = [1e8, -1e8]
 
-        obj = workspace.get_entity(objects.value)[0]
+        obj = workspace.get_entity(selection.objects.value)[0]
         if isinstance(obj, Grid2D):
 
-            objects.grid = obj
+            selection.objects.grid = obj
 
             lim_x[0], lim_x[1] = obj.centroids[:, 0].min(), obj.centroids[:, 0].max()
             lim_y[0], lim_y[1] = obj.centroids[:, 1].min(), obj.centroids[:, 1].max()
@@ -778,13 +773,13 @@ def edge_detection_widget(
                 center_y.value = numpy.mean(lim_y)
                 width_y.value = (lim_y[1] - lim_y[0]) / 2
 
-    objects.observe(set_bounding_box)
+    selection.objects.observe(set_bounding_box)
     set_bounding_box("", update_values=True)
 
     def saveIt(_):
         if export.value and getattr(export, "vertices", None) is not None:
             Curve.create(
-                objects.grid.workspace,
+                selection.objects.grid.workspace,
                 name=export_as.value,
                 vertices=export.vertices,
                 cells=export.cells,
@@ -796,12 +791,16 @@ def edge_detection_widget(
     export.observe(saveIt)
 
     def compute_trigger(_):
-        if compute.value and getattr(objects, "grid", None) is not None:
-            x = objects.grid.centroids[:, 0].reshape(objects.grid.shape, order="F")
-            y = objects.grid.centroids[:, 1].reshape(objects.grid.shape, order="F")
+        if compute.value and getattr(selection.objects, "grid", None) is not None:
+            x = selection.objects.grid.centroids[:, 0].reshape(
+                selection.objects.grid.shape, order="F"
+            )
+            y = selection.objects.grid.centroids[:, 1].reshape(
+                selection.objects.grid.shape, order="F"
+            )
 
-            grid_data = objects.grid.get_data(data_obj.value)[0].values
-            grid_data = grid_data.reshape(objects.grid.shape, order="F")
+            grid_data = selection.objects.grid.get_data(selection.data.value)[0].values
+            grid_data = grid_data.reshape(selection.objects.grid.shape, order="F")
 
             filter = filter_xy(x, y, resolution.value)
 
@@ -887,12 +886,12 @@ def edge_detection_widget(
 
                 if coords:
                     coord = numpy.vstack(coords)
-                    objects.lines = numpy.unique(coord, axis=1)
+                    selection.objects.lines = numpy.unique(coord, axis=1)
 
                     export.button_style = "success"
 
                 else:
-                    objects.lines = None
+                    selection.objects.lines = None
 
             compute.value = False
 
@@ -910,7 +909,7 @@ def edge_detection_widget(
         compute.button_style = "warning"
         export.button_style = "danger"
 
-    data_obj.observe(parameters_change)
+    selection.data.observe(parameters_change)
     resolution.observe(parameters_change)
     threshold.observe(parameters_change)
     sigma.observe(parameters_change)
@@ -929,10 +928,10 @@ def edge_detection_widget(
         compute,
     ):
 
-        if getattr(objects, "grid", None) is not None and objects.grid.get_data(
-            data_name
-        ):
-            data_obj = objects.grid.get_data(data_name)[0]
+        if getattr(
+            selection.objects, "grid", None
+        ) is not None and selection.objects.grid.get_data(data_name):
+            data_obj = selection.objects.grid.get_data(data_name)[0]
             fig = plt.figure(figsize=(10, 10))
             ax1 = plt.subplot()
 
@@ -948,8 +947,8 @@ def edge_detection_widget(
             corners = rotate_xy(corners, [0, 0], -azimuth)
             ax1.plot(corners[:, 0] + center_x, corners[:, 1] + center_y, "k")
 
-            _, _, ind_filter, _ = plot_plan_data_selection(
-                objects.grid,
+            _, _, ind_filter, _, _ = plot_plan_data_selection(
+                selection.objects.grid,
                 data_obj,
                 **{
                     "ax": ax1,
@@ -964,8 +963,8 @@ def edge_detection_widget(
                 },
             )
             data_count.value = f"Data Count: {ind_filter.sum()}"
-        if getattr(objects, "lines", None) is not None:
-            xy = objects.lines
+        if getattr(selection.objects, "lines", None) is not None:
+            xy = selection.objects.lines
 
             indices_1 = filter_xy(
                 xy[1::2, 0],
@@ -992,7 +991,7 @@ def edge_detection_widget(
                 numpy.any(numpy.c_[indices_1, indices_2], axis=1), numpy.ones(2),
             ).astype(bool)
 
-            xy = objects.lines[indices, :]
+            xy = selection.objects.lines[indices, :]
 
             ax1.add_collection(
                 collections.LineCollection(
@@ -1016,7 +1015,7 @@ def edge_detection_widget(
     plot_window = widgets.interactive_output(
         compute_plot,
         {
-            "data_name": data_obj,
+            "data_name": selection.data,
             "resolution": resolution,
             "center_x": center_x,
             "center_y": center_y,
@@ -1034,8 +1033,7 @@ def edge_detection_widget(
                 [
                     VBox(
                         [
-                            objects,
-                            data_obj,
+                            selection.widget,
                             resolution,
                             compute,
                             HBox([export, export_as]),
@@ -1065,1269 +1063,3 @@ def edge_detection_widget(
     )
 
     return out
-
-
-class EMLineProfiling:
-    groups = {
-        "early": {
-            "color": "blue",
-            "channels": [],
-            "gates": [],
-            "defaults": ["early"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-        "early + middle": {
-            "color": "cyan",
-            "channels": [],
-            "gates": [],
-            "defaults": ["early", "middle"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-        "middle": {
-            "color": "green",
-            "channels": [],
-            "gates": [],
-            "defaults": ["middle"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-        "early + middle + late": {
-            "color": "orange",
-            "channels": [],
-            "gates": [],
-            "defaults": ["early", "middle", "late"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-        "middle + late": {
-            "color": "yellow",
-            "channels": [],
-            "gates": [],
-            "defaults": ["middle", "late"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-        "late": {
-            "color": "red",
-            "channels": [],
-            "gates": [],
-            "defaults": ["late"],
-            "inflx_up": [],
-            "inflx_dwn": [],
-            "peaks": [],
-            "times": [],
-            "values": [],
-            "locations": [],
-        },
-    }
-
-    viz_param = '<IParameterList Version="1.0">\n <Colour>4278190335</Colour>\n <Transparency>0</Transparency>\n <Nodesize>9</Nodesize>\n <Nodesymbol>Sphere</Nodesymbol>\n <Scalenodesbydata>false</Scalenodesbydata>\n <Data></Data>\n <Scale>1</Scale>\n <Scalebyabsolutevalue>[NO STRING REPRESENTATION]</Scalebyabsolutevalue>\n <Orientation toggled="true">{\n    "DataGroup": "AzmDip",\n    "ManualWidth": true,\n    "Scale": false,\n    "ScaleLog": false,\n    "Size": 30,\n    "Symbol": "Tablet",\n    "Width": 30\n}\n</Orientation>\n</IParameterList>\n'
-
-    def __init__(self, h5file):
-
-        self.workspace = Workspace(h5file)
-        self.em_system_specs = utils.geophysical_systems.parameters()
-        self.system = widgets.Dropdown(
-            options=[
-                key
-                for key, specs in self.em_system_specs.items()
-                if specs["type"] == "time"
-            ],
-            description="Time-Domain System:",
-            style={"description_width": "initial"},
-        )
-
-        self._groups = self.groups
-        self.group_list = widgets.SelectMultiple(description="")
-
-        self.early = numpy.arange(8, 17).tolist()
-        self.middle = numpy.arange(17, 28).tolist()
-        self.late = numpy.arange(28, 40).tolist()
-
-        self.data_objects, self.data_field = object_data_selection_widget(
-            h5file, select_multiple=True
-        )
-        self.data_objects.description = "Survey"
-
-        self.model_objects, self.model_field = object_data_selection_widget(h5file)
-        self.model_objects.description = "1D Object:"
-        self.model_field.description = "Model"
-
-        _, self.model_line_field = object_data_selection_widget(
-            h5file, objects=self.model_objects, find_value=["line"]
-        )
-        self.model_line_field.description = "Line field: "
-
-        self.marker = {"left": "<", "right": ">"}
-
-        def update_model_line_fields(_):
-            self.model_line_field.options = self.model_field.options
-            self.model_line_field.value = find_value(
-                self.model_line_field.options, ["line"]
-            )
-
-        self.model_field.observe(update_model_line_fields, names="options")
-
-        def get_survey(_):
-            if self.workspace.get_entity(self.data_objects.value):
-                self.survey = self.workspace.get_entity(self.data_objects.value)[0]
-                self.data_field.options = (
-                    [p_g.name for p_g in self.survey.property_groups]
-                    + ["^-- Groups --^"]
-                    + list(self.data_field.options)
-                )
-
-        self.data_objects.observe(get_survey, names="value")
-        self.data_objects.value = self.data_objects.options[0]
-        get_survey("")
-
-        self.lines = LineOptions(h5file, self.data_objects, select_multiple=False)
-        self.lines.value.description = "Line"
-
-        self.channels = widgets.SelectMultiple(description="Channels")
-        self.group_default_early = widgets.Text(description="Early", value="9-16")
-        self.group_default_middle = widgets.Text(description="Middle", value="17-27")
-        self.group_default_late = widgets.Text(description="Late", value="28-40")
-
-        def reset_default_bounds(_):
-            self.reset_default_bounds()
-
-        self.data = {}
-        self.data_channel_options = {}
-
-        def get_data(_):
-            data = []
-
-            groups = [p_g.name for p_g in self.survey.property_groups]
-            channels = list(self.data_field.value)
-
-            for channel in self.data_field.value:
-                if channel in groups:
-                    for prop in self.survey.get_property_group(channel).properties:
-                        name = self.workspace.get_entity(prop)[0].name
-                        if prop not in channels:
-                            channels.append(name)
-
-            self.channels.options = channels
-            for channel in channels:
-                if self.survey.get_data(channel):
-                    self.data[channel] = self.survey.get_data(channel)[0]
-
-            # Generate default groups
-            self.reset_default_bounds()
-
-            for key, widget in self.data_channel_options.items():
-                widget.children[0].options = channels
-                widget.children[0].value = find_value(channels, [key])
-
-        self.data_field.observe(get_data, names="value")
-        get_data("")
-
-        def get_surf_model(_):
-            if self.workspace.get_entity(self.model_objects.value):
-                self.surf_model = self.workspace.get_entity(self.model_objects.value)[0]
-
-        self.model_objects.observe(get_surf_model, names="value")
-
-        def get_model(_):
-            if self.surf_model.get_data(self.model_field.value):
-                self.surf_model = self.surf_model.get_data(self.model_field.value)[0]
-
-        self.model_objects.observe(get_model, names="value")
-
-        self.smoothing = widgets.IntSlider(
-            min=0,
-            max=64,
-            value=0,
-            description="Running mean",
-            continuous_update=False,
-            style={"description_width": "initial"},
-        )
-
-        self.residual = widgets.Checkbox(description="Use residual", value=False)
-
-        self.threshold = widgets.FloatSlider(
-            value=50,
-            min=10,
-            max=90,
-            step=5,
-            continuous_update=False,
-            description="Decay threshold (%)",
-        )
-
-        self.center = widgets.FloatSlider(
-            value=0.5,
-            min=0,
-            max=1.0,
-            step=0.001,
-            description="Center (%)",
-            disabled=False,
-            continuous_update=False,
-            orientation="horizontal",
-            style={"description_width": "initial"},
-        )
-
-        self.auto_picker = widgets.ToggleButton(description="Pick target", value=True)
-
-        self.focus = widgets.FloatSlider(
-            value=1.0,
-            min=0.025,
-            max=1.0,
-            step=0.005,
-            description="Width (%)",
-            disabled=False,
-            continuous_update=False,
-            orientation="horizontal",
-        )
-
-        self.zoom = widgets.VBox([self.center, self.focus])
-
-        self.scale_button = widgets.RadioButtons(
-            options=["linear", "symlog",],
-            description="Vertical scaling",
-            style={"description_width": "initial"},
-        )
-
-        def scale_update(_):
-            if self.scale_button.value == "symlog":
-                scale_panel.children = [
-                    self.scale_button,
-                    widgets.VBox([widgets.Label("Linear threshold"), self.scale_value]),
-                ]
-            else:
-                scale_panel.children = [self.scale_button]
-
-        self.scale_button.observe(scale_update)
-        self.scale_value = widgets.FloatLogSlider(
-            min=-18,
-            max=10,
-            step=0.5,
-            base=10,
-            value=1e-2,
-            description="",
-            continuous_update=False,
-            style={"description_width": "initial"},
-        )
-        scale_panel = widgets.HBox([self.scale_button])
-
-        def add_group(_):
-            self.add_group()
-
-        def channel_setter(caller):
-
-            channel = caller["owner"]
-            data_widget = self.data_channel_options[channel.header]
-            data_widget.children[0].value = find_value(
-                data_widget.children[0].options, [channel.header]
-            )
-
-        self.channel_selection = widgets.Dropdown(
-            description="Time Gate",
-            style={"description_width": "initial"},
-            options=self.em_system_specs[self.system.value]["channels"].keys(),
-        )
-
-        def system_observer(_):
-
-            system_specs = {}
-            for key, time_gate in self.em_system_specs[self.system.value][
-                "channels"
-            ].items():
-                system_specs[key] = f"{time_gate:.5e}"
-
-            self.channel_selection.options = self.em_system_specs[self.system.value][
-                "channels"
-            ].keys()
-
-            self.data_channel_options = {}
-            for ind, (key, value) in enumerate(system_specs.items()):
-                channel_selection = widgets.Dropdown(
-                    description="Channel",
-                    style={"description_width": "initial"},
-                    options=self.channels.options,
-                    value=find_value(self.channels.options, [key]),
-                )
-                channel_selection.header = key
-                channel_selection.observe(channel_setter, names="value")
-
-                channel_time = widgets.FloatText(description="Time (s)", value=value)
-
-                self.data_channel_options[key] = widgets.VBox(
-                    [channel_selection, channel_time]
-                )
-
-        self.system.observe(system_observer)
-        system_observer("")
-
-        def channel_panel_update(_):
-            self.channel_panel.children = [
-                self.channel_selection,
-                self.data_channel_options[self.channel_selection.value],
-            ]
-
-        self.channel_selection.observe(channel_panel_update, names="value")
-        self.channel_panel = widgets.VBox(
-            [
-                self.channel_selection,
-                self.data_channel_options[self.channel_selection.value],
-            ]
-        )
-
-        self.group_default_early.observe(reset_default_bounds)
-        self.group_default_middle.observe(reset_default_bounds)
-        self.group_default_late.observe(reset_default_bounds)
-
-        self.group_add = widgets.ToggleButton(description="<< Add New Group <<")
-        self.group_name = widgets.Text(description="Name")
-        self.group_color = widgets.ColorPicker(
-            concise=False, description="Color", value="blue", disabled=False
-        )
-        self.group_add.observe(add_group)
-
-        def highlight_selection(_):
-
-            self.highlight_selection()
-
-        self.group_list.observe(highlight_selection, names="value")
-        self.markers = widgets.ToggleButton(description="Show all markers")
-
-        def export_click(_):
-            self.export_click()
-
-        self.export = widgets.ToggleButton(description="Export", button_style="success")
-        self.export.observe(export_click)
-
-        def plot_data_selection(
-            ind,
-            smoothing,
-            residual,
-            markers,
-            scale,
-            scale_value,
-            center,
-            focus,
-            groups,
-            pick_trigger,
-            x_label,
-            threshold,
-        ):
-            self.plot_data_selection(
-                ind,
-                smoothing,
-                residual,
-                markers,
-                scale,
-                scale_value,
-                center,
-                focus,
-                groups,
-                pick_trigger,
-                x_label,
-                threshold,
-            )
-
-        def plot_model_selection(ind, center, focus):
-            self.plot_model_selection(ind, center, focus)
-
-        self.x_label = widgets.ToggleButtons(
-            options=["Distance", "Easting", "Northing"],
-            value="Distance",
-            description="X-axis label:",
-        )
-        plotting = widgets.interactive_output(
-            plot_data_selection,
-            {
-                "ind": self.lines.lines,
-                "smoothing": self.smoothing,
-                "residual": self.residual,
-                "markers": self.markers,
-                "scale": self.scale_button,
-                "scale_value": self.scale_value,
-                "center": self.center,
-                "focus": self.focus,
-                "groups": self.group_list,
-                "pick_trigger": self.auto_picker,
-                "x_label": self.x_label,
-                "threshold": self.threshold,
-            },
-        )
-
-        section = widgets.interactive_output(
-            plot_model_selection,
-            {"ind": self.lines.lines, "center": self.center, "focus": self.focus,},
-        )
-
-        self.model_panel = widgets.VBox(
-            [
-                widgets.HBox(
-                    [
-                        self.model_objects,
-                        widgets.VBox([self.model_field, self.model_line_field]),
-                    ]
-                ),
-                section,
-            ]
-        )
-
-        self.show_model = widgets.Checkbox(description="Show model", value=False)
-
-        def show_model_trigger(_):
-            self.show_model_trigger()
-
-        self.show_model.observe(show_model_trigger)
-
-        self.live_link = widgets.Checkbox(
-            description="GA Pro - Live link", value=False, indent=False
-        )
-
-        def live_link_trigger(_):
-            self.live_link_trigger()
-
-        self.live_link.observe(live_link_trigger)
-
-        self.live_link_path = widgets.Text(
-            description="",
-            value=os.path.join(os.path.dirname(h5file), "Temp"),
-            disabled=True,
-            style={"description_width": "initial"},
-        )
-
-        self.data_panel = widgets.VBox(
-            [
-                widgets.HBox(
-                    [
-                        widgets.VBox([self.data_objects, self.data_field]),
-                        widgets.VBox([self.system, self.channel_panel]),
-                    ]
-                ),
-                self.lines.widget,
-                plotting,
-                self.x_label,
-                widgets.HBox(
-                    [
-                        widgets.VBox(
-                            [
-                                widgets.HBox(
-                                    [
-                                        widgets.VBox(
-                                            [self.zoom, scale_panel],
-                                            layout=widgets.Layout(width="50%"),
-                                        ),
-                                        widgets.VBox(
-                                            [
-                                                self.smoothing,
-                                                self.residual,
-                                                self.threshold,
-                                            ],
-                                            layout=widgets.Layout(width="50%"),
-                                        ),
-                                    ]
-                                ),
-                                widgets.HBox([self.markers, self.auto_picker]),
-                                widgets.VBox(
-                                    [
-                                        widgets.Label("Groups"),
-                                        widgets.HBox(
-                                            [
-                                                widgets.Label("Defaults"),
-                                                self.group_default_early,
-                                                self.group_default_middle,
-                                                self.group_default_late,
-                                            ]
-                                        ),
-                                        widgets.HBox(
-                                            [
-                                                self.group_list,
-                                                widgets.VBox(
-                                                    [
-                                                        self.channels,
-                                                        self.group_name,
-                                                        self.group_color,
-                                                        self.group_add,
-                                                    ]
-                                                ),
-                                            ]
-                                        ),
-                                        widgets.HBox(
-                                            [
-                                                self.export,
-                                                widgets.VBox(
-                                                    [
-                                                        self.live_link,
-                                                        widgets.Label(
-                                                            "Monitoring folder"
-                                                        ),
-                                                        self.live_link_path,
-                                                    ]
-                                                ),
-                                            ]
-                                        ),
-                                    ]
-                                ),
-                            ]
-                        )
-                    ]
-                ),
-            ]
-        )
-
-        self._widget = widgets.VBox([self.data_panel, self.show_model])
-
-    def set_default_groups(self, channels):
-        """
-        Assign TEM channel for given gate #
-        """
-        # Reset channels
-        for group in self.groups.values():
-            if len(group["defaults"]) > 0:
-                group["channels"] = []
-                group["inflx_up"] = []
-                group["inflx_dwn"] = []
-                group["peaks"] = []
-                group["mad_tau"] = []
-                group["times"] = []
-                group["values"] = []
-                group["locations"] = []
-
-        for ind, channel in enumerate(channels):
-            for group in self.groups.values():
-                if ind in group["gates"]:
-                    group["channels"].append(channel)
-
-        self.group_list.options = self.groups.keys()
-        self.group_list.value = []
-
-    def add_group(self):
-        """
-        Add a group to the list of groups
-        """
-        if self.group_add.value:
-
-            if self.group_name.value not in self.group_list.options:
-                self.group_list.options = list(self.group_list.options) + [
-                    self.group_name.value
-                ]
-
-            self.groups[self.group_name.value] = {
-                "color": self.group_color.value,
-                "channels": list(self.channels.value),
-                "inflx_up": [],
-                "inflx_dwn": [],
-                "peaks": [],
-                "mad_tau": [],
-                "times": [],
-                "values": [],
-                "locations": [],
-            }
-            self.group_add.value = False
-
-    def export_click(self):
-        if self.export.value:
-            for group in self.group_list.value:
-
-                for (
-                    ind,
-                    (channel, locations, peaks, inflx_dwn, inflx_up, vals, times),
-                ) in enumerate(
-                    zip(
-                        self.groups[group]["channels"],
-                        self.groups[group]["locations"],
-                        self.groups[group]["peaks"],
-                        self.groups[group]["inflx_dwn"],
-                        self.groups[group]["inflx_up"],
-                        self.groups[group]["values"],
-                        self.groups[group]["times"],
-                    )
-                ):
-
-                    if ind == 0:
-                        cox_x = self.lines.profile.interp_x(peaks[0])
-                        cox_y = self.lines.profile.interp_y(peaks[0])
-                        cox_z = self.lines.profile.interp_z(peaks[0])
-                        cox = numpy.r_[cox_x, cox_y, cox_z]
-
-                        # Compute average dip
-                        left_ratio = numpy.abs(
-                            (peaks[1] - inflx_up[1]) / (peaks[0] - inflx_up[0])
-                        )
-                        right_ratio = numpy.abs(
-                            (peaks[1] - inflx_dwn[1]) / (peaks[0] - inflx_dwn[0])
-                        )
-
-                        if left_ratio > right_ratio:
-                            ratio = right_ratio / left_ratio
-                            azm = (
-                                450.0
-                                - numpy.rad2deg(
-                                    numpy.arctan2(
-                                        (
-                                            self.lines.profile.interp_y(inflx_up[0])
-                                            - cox_y
-                                        ),
-                                        (
-                                            self.lines.profile.interp_x(inflx_up[0])
-                                            - cox_x
-                                        ),
-                                    )
-                                )
-                            ) % 360.0
-                        else:
-                            ratio = left_ratio / right_ratio
-                            azm = (
-                                450.0
-                                - numpy.rad2deg(
-                                    numpy.arctan2(
-                                        (
-                                            self.lines.profile.interp_y(inflx_dwn[0])
-                                            - cox_y
-                                        ),
-                                        (
-                                            self.lines.profile.interp_x(inflx_dwn[0])
-                                            - cox_x
-                                        ),
-                                    )
-                                )
-                            ) % 360.0
-
-                        dip = numpy.rad2deg(numpy.arcsin(ratio))
-                    tau = self.groups[group]["mad_tau"]
-
-                if self.workspace.get_entity(group):
-                    points = self.workspace.get_entity(group)[0]
-                    azm_data = points.get_data("azimuth")[0]
-                    azm_vals = azm_data.values.copy()
-                    dip_data = points.get_data("dip")[0]
-                    dip_vals = dip_data.values.copy()
-
-                    tau_data = points.get_data("tau")[0]
-                    tau_vals = tau_data.values.copy()
-
-                    points.vertices = numpy.vstack(
-                        [points.vertices, cox.reshape((1, 3))]
-                    )
-                    azm_data.values = numpy.hstack([azm_vals, azm])
-                    dip_data.values = numpy.hstack([dip_vals, dip])
-                    tau_data.values = numpy.hstack([tau_vals, tau])
-
-                else:
-                    # if self.workspace.get_entity(group)
-                    # parent =
-                    points = Points.create(
-                        self.workspace, name=group, vertices=cox.reshape((1, 3))
-                    )
-                    points.add_data(
-                        {
-                            "azimuth": {"values": numpy.asarray(azm)},
-                            "dip": {"values": numpy.asarray(dip)},
-                            "tau": {"values": numpy.asarray(tau)},
-                            # "Visual Parameters": {"values": self.viz_param}
-                        }
-                    )
-                    group = points.find_or_create_property_group(
-                        name="AzmDip", property_group_type="Dip direction & dip"
-                    )
-                    group.properties = [
-                        points.get_data("azimuth")[0].uid,
-                        points.get_data("dip")[0].uid,
-                    ]
-
-                if self.live_link.value:
-                    if not os.path.exists(self.live_link_path.value):
-                        os.mkdir(self.live_link_path.value)
-
-                    temp_geoh5 = os.path.join(
-                        self.live_link_path.value, f"temp{time.time():.3f}.geoh5"
-                    )
-                    ws_out = Workspace(temp_geoh5)
-                    points.copy(parent=ws_out)
-
-            self.export.value = False
-            self.workspace.finalize()
-
-    def highlight_selection(self):
-        """
-        Highlight the group choice
-        """
-        highlights = []
-        for group in self.group_list.value:
-            highlights += self.groups[group]["channels"]
-            self.group_color.value = self.groups[group]["color"]
-        self.channels.value = highlights
-
-    def live_link_trigger(self):
-        """
-        Enable the monitoring folder
-        """
-        if self.live_link.value:
-            self.live_link_path.disabled = False
-        else:
-            self.live_link_path.disabled = True
-
-    def plot_data_selection(
-        self,
-        ind,
-        smoothing,
-        residual,
-        markers,
-        scale,
-        scale_value,
-        center,
-        focus,
-        groups,
-        pick_trigger,
-        x_label,
-        threshold,
-    ):
-
-        fig = plt.figure(figsize=(12, 8))
-        ax2 = plt.subplot()
-
-        self.line_update()
-
-        for group in self.groups.values():
-            group["inflx_up"] = []
-            group["inflx_dwn"] = []
-            group["peaks"] = []
-            group["mad_tau"]
-            group["times"] = []
-            group["values"] = []
-            group["locations"] = []
-
-        if (
-            getattr(self, "survey", None) is None
-            or getattr(self.lines, "profile", None) is None
-        ):
-            return
-
-        center_x = center * self.lines.profile.locations_resampled[-1]
-
-        if residual != self.lines.profile.residual:
-            self.lines.profile.residual = residual
-            self.line_update()
-
-        if smoothing != self.lines.profile.smoothing:
-            self.lines.profile.smoothing = smoothing
-            self.line_update()
-
-        lims = numpy.searchsorted(
-            self.lines.profile.locations_resampled,
-            [
-                (center - focus / 2.0) * self.lines.profile.locations_resampled[-1],
-                (center + focus / 2.0) * self.lines.profile.locations_resampled[-1],
-            ],
-        )
-
-        sub_ind = numpy.arange(lims[0], lims[1])
-
-        channels = []
-        for group in self.group_list.value:
-            channels += self.groups[group]["channels"]
-
-        if len(channels) == 0:
-            channels = self.channels.options
-
-        times = {}
-        for channel in self.data_channel_options.values():
-            times[channel.children[0].value] = channel.children[1].value
-
-        y_min, y_max = numpy.inf, -numpy.inf
-        for channel, d in self.data.items():
-
-            if channel not in times.keys():
-                continue
-
-            if channel not in channels:
-                continue
-
-            self.lines.profile.values = d.values[self.survey.line_indices].copy()
-            locs, values = (
-                self.lines.profile.locations_resampled[sub_ind],
-                self.lines.profile.values_resampled[sub_ind],
-            )
-
-            y_min = numpy.min([values.min(), y_min])
-            y_max = numpy.max([values.max(), y_max])
-
-            ax2.plot(locs, values, color=[0.5, 0.5, 0.5, 1])
-
-            if not residual:
-                raw = self.lines.profile._values_resampled_raw[sub_ind]
-                ax2.fill_between(
-                    locs, values, raw, where=raw > values, color=[1, 0, 0, 0.5]
-                )
-                ax2.fill_between(
-                    locs, values, raw, where=raw < values, color=[0, 0, 1, 0.5]
-                )
-
-            dx = self.lines.profile.derivative(order=1)[sub_ind]
-            ddx = self.lines.profile.derivative(order=2)[sub_ind]
-
-            peaks = numpy.where((numpy.diff(numpy.sign(dx)) != 0) * (ddx[1:] < 0))[0]
-            lows = numpy.where((numpy.diff(numpy.sign(dx)) != 0) * (ddx[1:] > 0))[0]
-
-            up_inflx = numpy.where((numpy.diff(numpy.sign(ddx)) != 0) * (dx[1:] > 0))[0]
-            dwn_inflx = numpy.where((numpy.diff(numpy.sign(ddx)) != 0) * (dx[1:] < 0))[
-                0
-            ]
-
-            if markers:
-                ax2.scatter(locs[peaks], values[peaks], s=100, color="r", marker="v")
-                ax2.scatter(locs[lows], values[lows], s=100, color="b", marker="^")
-                ax2.scatter(locs[up_inflx], values[up_inflx], color="g")
-                ax2.scatter(locs[dwn_inflx], values[dwn_inflx], color="m")
-
-            if len(peaks) == 0 or len(lows) < 2:
-                continue
-
-            if self.auto_picker.value:
-
-                # Check dipping direction
-                x_ind = numpy.min(
-                    [values.shape[0] - 2, numpy.searchsorted(locs, center_x)]
-                )
-                if values[x_ind] > values[x_ind + 1]:
-                    cox = peaks[numpy.searchsorted(locs[peaks], center_x) - 1]
-                else:
-                    x_ind = numpy.min(
-                        [peaks.shape[0] - 1, numpy.searchsorted(locs[peaks], center_x)]
-                    )
-                    cox = peaks[x_ind]
-
-                start = lows[numpy.searchsorted(locs[lows], locs[cox]) - 1]
-                end = lows[
-                    numpy.min(
-                        [numpy.searchsorted(locs[lows], locs[cox]), len(lows) - 1]
-                    )
-                ]
-
-                bump_x = locs[start:end]
-                bump_v = values[start:end]
-
-                if len(bump_x) == 0:
-                    continue
-
-                inflx_up = numpy.searchsorted(
-                    bump_x,
-                    locs[up_inflx[numpy.searchsorted(locs[up_inflx], locs[cox]) - 1]],
-                )
-                inflx_up = numpy.max([0, inflx_up])
-
-                inflx_dwn = numpy.searchsorted(
-                    bump_x,
-                    locs[dwn_inflx[numpy.searchsorted(locs[dwn_inflx], locs[cox])]],
-                )
-                inflx_dwn = numpy.min([bump_x.shape[0] - 1, inflx_dwn])
-
-                peak = numpy.min(
-                    [bump_x.shape[0] - 1, numpy.searchsorted(bump_x, locs[cox])]
-                )
-
-                for ii, group in enumerate(self.group_list.value):
-                    if channel in self.groups[group]["channels"]:
-                        self.groups[group]["inflx_up"].append(
-                            numpy.r_[bump_x[inflx_up], bump_v[inflx_up]]
-                        )
-                        self.groups[group]["peaks"].append(
-                            numpy.r_[bump_x[peak], bump_v[peak]]
-                        )
-                        self.groups[group]["times"].append(times[channel])
-                        self.groups[group]["inflx_dwn"].append(
-                            numpy.r_[bump_x[inflx_dwn], bump_v[inflx_dwn]]
-                        )
-                        self.groups[group]["locations"].append(bump_x)
-                        self.groups[group]["values"].append(bump_v)
-
-                        # Compute average dip
-                        left_ratio = (bump_v[peak] - bump_v[inflx_up]) / (
-                            bump_x[peak] - bump_x[inflx_up]
-                        )
-                        right_ratio = (bump_v[peak] - bump_v[inflx_dwn]) / (
-                            bump_x[inflx_dwn] - bump_x[peak]
-                        )
-
-                        if left_ratio > right_ratio:
-                            ratio = right_ratio / left_ratio
-                            ori = "left"
-                        else:
-                            ratio = left_ratio / right_ratio
-                            ori = "right"
-
-                        dip = numpy.rad2deg(numpy.arcsin(ratio))
-
-                        # Left
-                        ax2.plot(
-                            bump_x[:peak],
-                            bump_v[:peak],
-                            "--",
-                            color=self.groups[group]["color"],
-                        )
-                        # Right
-                        ax2.plot(
-                            bump_x[peak:],
-                            bump_v[peak:],
-                            color=self.groups[group]["color"],
-                        )
-                        ax2.scatter(
-                            self.groups[group]["peaks"][-1][0],
-                            self.groups[group]["peaks"][-1][1],
-                            s=100,
-                            c=self.groups[group]["color"],
-                            marker=self.marker[ori],
-                        )
-                        if ~numpy.isnan(dip):
-                            ax2.text(
-                                self.groups[group]["peaks"][-1][0],
-                                self.groups[group]["peaks"][-1][1],
-                                f"{dip:.0f}",
-                                va="bottom",
-                                ha="center",
-                            )
-                        ax2.scatter(
-                            self.groups[group]["inflx_dwn"][-1][0],
-                            self.groups[group]["inflx_dwn"][-1][1],
-                            s=100,
-                            c=self.groups[group]["color"],
-                            marker="1",
-                        )
-                        ax2.scatter(
-                            self.groups[group]["inflx_up"][-1][0],
-                            self.groups[group]["inflx_up"][-1][1],
-                            s=100,
-                            c=self.groups[group]["color"],
-                            marker="2",
-                        )
-
-        ax2.plot(
-            [
-                self.lines.profile.locations_resampled[0],
-                self.lines.profile.locations_resampled[-1],
-            ],
-            [0, 0],
-            "r",
-        )
-        ax2.plot([center_x, center_x], [0, y_min], "r--")
-        ax2.scatter(center_x, y_min, s=20, c="r", marker="^")
-
-        for group in self.groups.values():
-            if group["peaks"]:
-                peaks = numpy.vstack(group["peaks"])
-                ratio = peaks[:, 1] / peaks[0, 1]
-                ind = numpy.where(ratio >= (threshold / 100))[0][-1]
-                #                 print(ind)
-                ax2.plot(
-                    peaks[: ind + 1, 0], peaks[: ind + 1, 1], "--", color=group["color"]
-                )
-        #                 ax2.plot([peaks[0, 0], peaks[0, 0]], [peaks[0, 1], peaks[-1, 1]], '--', color='k')
-        #                 ax2.plot(
-        #                     [group['inflx_up'][ind][0], group['inflx_dwn'][ind][0]]
-        #                     [group['inflx_up'][ind][1], group['inflx_dwn'][ind][1]], '--', color=[0.5,0.5,0.5]
-        #                 )
-
-        if scale == "symlog":
-            plt.yscale("symlog", linthreshy=scale_value)
-
-        x_lims = [
-            center_x - focus / 2.0 * self.lines.profile.locations_resampled[-1],
-            center_x + focus / 2.0 * self.lines.profile.locations_resampled[-1],
-        ]
-        ax2.set_xlim(x_lims)
-        ax2.set_title(f"Line: {ind}")
-        ax2.set_ylabel("dBdT")
-
-        if x_label == "Easting":
-
-            ax2.text(
-                center_x,
-                0,
-                f"{self.lines.profile.interp_x(center_x):.0f} m E",
-                va="top",
-                ha="center",
-                bbox={"edgecolor": "r"},
-            )
-            xlbl = [
-                f"{self.lines.profile.interp_x(label):.0f}"
-                for label in ax2.get_xticks()
-            ]
-            ax2.set_xticklabels(xlbl)
-            ax2.set_xlabel("Easting (m)")
-        elif x_label == "Northing":
-            ax2.text(
-                center_x,
-                0,
-                f"{self.lines.profile.interp_y(center_x):.0f} m N",
-                va="top",
-                ha="center",
-                bbox={"edgecolor": "r"},
-            )
-            xlbl = [
-                f"{self.lines.profile.interp_y(label):.0f}"
-                for label in ax2.get_xticks()
-            ]
-            ax2.set_xticklabels(xlbl)
-            ax2.set_xlabel("Northing (m)")
-        else:
-            ax2.text(
-                center_x,
-                0,
-                f"{center_x:.0f} m",
-                va="top",
-                ha="center",
-                bbox={"edgecolor": "r"},
-            )
-            ax2.set_xlabel("Distance (m)")
-
-        ax2.grid(True)
-
-        pos2 = ax2.get_position()
-
-        ax = [pos2.x0, pos2.y0, pos2.width, pos2.height].copy()
-
-        if self.auto_picker.value:
-            ax[0] += 0.25
-            ax[1] -= 0.5
-            ax[2] /= 3
-            ax[3] /= 2
-            ax4 = plt.axes(ax)
-            for group in self.group_list.value:
-                if len(self.groups[group]["peaks"]) == 0:
-                    continue
-
-                peaks = (
-                    numpy.vstack(self.groups[group]["peaks"])
-                    * self.em_system_specs[self.system.value]["normalization"]
-                )
-
-                tc = numpy.hstack(self.groups[group]["times"][: ind + 1])
-                vals = numpy.log(peaks[: ind + 1, 1])
-
-                if tc.shape[0] < 2:
-                    continue
-                # Compute linear trend
-                A = numpy.c_[numpy.ones_like(tc), tc]
-                a, c = numpy.linalg.solve(numpy.dot(A.T, A), numpy.dot(A.T, vals))
-                d = numpy.r_[tc.min(), tc.max()]
-                vv = d * c + a
-
-                ratio = numpy.abs((vv[0] - vv[1]) / (d[0] - d[1]))
-                #                 angl = numpy.arctan(ratio**-1.)
-
-                self.groups[group]["mad_tau"] = ratio ** -1.0
-
-                ax4.plot(
-                    d,
-                    numpy.exp(d * c + a),
-                    "--",
-                    linewidth=2,
-                    color=self.groups[group]["color"],
-                )
-                ax4.text(
-                    numpy.mean(d),
-                    numpy.exp(numpy.mean(vv)),
-                    f"{ratio ** -1.:.2e}",
-                    color=self.groups[group]["color"],
-                )
-                #                 plt.yscale('symlog', linthreshy=scale_value)
-                #                 ax4.set_aspect('equal')
-                ax4.scatter(
-                    numpy.hstack(self.groups[group]["times"]),
-                    peaks[:, 1],
-                    color=self.groups[group]["color"],
-                    marker="^",
-                )
-                ax4.grid(True)
-
-            plt.yscale("symlog")
-            ax4.yaxis.set_label_position("right")
-            ax4.yaxis.tick_right()
-            ax4.set_ylabel("log(V)")
-            ax4.set_xlabel("Time (sec)")
-            ax4.set_title("Decay - MADTau")
-
-    def plot_model_selection(self, ind, center, focus):
-
-        fig = plt.figure(figsize=(12, 8))
-        ax3 = plt.subplot()
-
-        if (
-            getattr(self, "survey", None) is None
-            or getattr(self.lines, "profile", None) is None
-        ):
-            return
-
-        center_x = center * self.lines.profile.locations_resampled[-1]
-
-        x_lims = [
-            center_x - focus / 2.0 * self.lines.profile.locations_resampled[-1],
-            center_x + focus / 2.0 * self.lines.profile.locations_resampled[-1],
-        ]
-
-        if getattr(self.lines, "model_x", None) is not None:
-            return
-
-            cs = ax3.tricontourf(
-                self.lines.model_x,
-                self.lines.model_z,
-                self.lines.model_cells.reshape((-1, 3)),
-                numpy.log10(self.lines.model_values),
-                levels=numpy.linspace(-3, 0.5, 25),
-                vmin=-2,
-                vmax=-0.75,
-                cmap="rainbow",
-            )
-            ax3.tricontour(
-                self.lines.model_x,
-                self.lines.model_z,
-                self.lines.model_cells.reshape((-1, 3)),
-                numpy.log10(self.lines.model_values),
-                levels=numpy.linspace(-3, 0.5, 25),
-                colors="k",
-                linestyles="solid",
-                linewidths=0.5,
-            )
-            #         ax3.scatter(center_x, center_z, 100, c='r', marker='x')
-            ax3.set_xlim(x_lims)
-            ax3.set_aspect("equal")
-            ax3.grid(True)
-
-    def line_update(self):
-        """
-        Re-compute derivatives
-        """
-
-        if getattr(self, "survey", None) is None:
-            return
-
-        if (
-            len(self.survey.get_data(self.lines.value.value)) == 0
-            or self.lines.lines.value == ""
-        ):
-            return
-
-        line_ind = numpy.where(
-            numpy.asarray(self.survey.get_data(self.lines.value.value)[0].values)
-            == self.lines.lines.value
-        )[0]
-
-        if len(line_ind) == 0:
-            return
-
-        self.survey.line_indices = line_ind
-        xyz = self.survey.vertices[line_ind, :]
-
-        if numpy.std(xyz[:, 1]) > numpy.std(xyz[:, 0]):
-            start = numpy.argmin(xyz[:, 1])
-        else:
-            start = numpy.argmin(xyz[:, 0])
-
-        self.lines.profile = signal_processing_1d(
-            xyz, None, smoothing=self.smoothing.value, residual=self.residual.value
-        )
-
-        # Get the corresponding along line model
-        origin = xyz[0, :2]
-
-        if self.workspace.get_entity(self.model_objects.value):
-            surf_model = self.workspace.get_entity(self.model_objects.value)[0]
-
-        if surf_model.get_data("Line") and numpy.any(
-            numpy.where(
-                surf_model.get_data("Line")[0].values == self.lines.lines.value
-            )[0]
-        ):
-
-            surf_id = surf_model.get_data("Line")[0].values
-            #             surf_ind = numpy.where(
-            #                 surf_id == self.lines.lines.value
-            #             )[0]
-
-            cell_ind = numpy.where(
-                surf_id[surf_model.cells[:, 0]] == self.lines.lines.value
-            )[0]
-
-            cells = surf_model.cells[cell_ind, :]
-            vert_ind, cell_ind = numpy.unique(cells, return_inverse=True)
-
-            surf_verts = surf_model.vertices[vert_ind, :]
-            self.lines.model_x = numpy.linalg.norm(
-                numpy.c_[
-                    xyz[start, 0] - surf_verts[:, 0], xyz[start, 1] - surf_verts[:, 1]
-                ],
-                axis=1,
-            )
-            self.lines.model_z = surf_model.vertices[vert_ind, 2]
-            self.lines.model_cells = cell_ind
-            self.lines.model_values = surf_model.get_data(self.model_field.value)[
-                0
-            ].values[vert_ind]
-        else:
-            self.lines.model_x = None
-
-    def reset_default_bounds(self):
-
-        try:
-            first, last = numpy.asarray(
-                self.group_default_early.value.split("-"), dtype="int"
-            )
-            self.early = numpy.arange(first, last + 1).tolist()
-        except ValueError:
-            return
-
-        try:
-            first, last = numpy.asarray(
-                self.group_default_middle.value.split("-"), dtype="int"
-            )
-            self.middle = numpy.arange(first, last + 1).tolist()
-        except ValueError:
-            return
-
-        try:
-            first, last = numpy.asarray(
-                self.group_default_late.value.split("-"), dtype="int"
-            )
-            self.last = numpy.arange(first, last + 1).tolist()
-        except ValueError:
-            return
-
-        for group in self.groups.values():
-            gates = []
-            if len(group["defaults"]) > 0:
-                for default in group["defaults"]:
-                    gates += getattr(self, default)
-                group["gates"] = gates
-
-        self.set_default_groups(self.channels.options)
-
-    def show_model_trigger(self):
-        """
-        Add the model widget
-        """
-        if self.show_model.value:
-            self._widget.children = [self.data_panel, self.show_model, self.model_panel]
-        else:
-            self._widget.children = [self.data_panel, self.show_model]
-
-    @property
-    def widget(self):
-        return self._widget
