+++ geoapps/simulation.py
@@ -7,7 +7,7 @@ from geoh5py.groups import ContainerGroup
 from geoh5py.objects import Surface
 from geoh5py.workspace import Workspace
 
-from geoapps.selection import object_data_selection_widget
+from geoapps.selection import ObjectDataSelection
 from geoapps.simpegPF import PF, Maps, Utils
 from geoapps.utils import RectangularBlock, tensor_2_block_model
 
@@ -104,7 +104,7 @@ def block_model_widget(h5file, inducing_field="50000, 90, 0"):
 
         set_axes_equal(axs)
 
-    object_selection = object_data_selection_widget(h5file, interactive=True)
+    object_selection = ObjectDataSelection(h5file, interactive=True).widget
 
     obj = workspace.get_entity(object_selection.children[0].children[0].value)[0]
     if getattr(obj, "vertices", None) is not None:
