+++ geoapps/utils/__init__.py
@@ -16,4 +16,5 @@ from .utils import (
     data_2_zarr,
     rotate_vertices,
     RectangularBlock,
+    format_labels,
 )
