+++ readthedocs.yml
@@ -6,4 +6,3 @@ sphinx:
   builder: html
   configuration: conf.py
   fail_on_warning: true
-formats: []
